Back to only one design document, one view returning all data
The view returns the following array:
[<maximum value of session>, <user results>, <num user answers>]

However, when a user's individual results are needed, the session
maximum is not available. As a result, two queries are needed.
On the other hand, we usually want to show results for 'me' as
well as for the whole session at the same time, so we can return
both results on the server side.